Refactor: Move dodgems logic to Vehicle.Dodgems.cpp
Moves dodgems mode, motion, and collision detection to a dedicated file
